+++ src/main/g8/app/models/package.scala
@@ -77,7 +77,6 @@ package object models {
           val updatedJsArray = valueToRemoveFrom.value.slice(0, index) ++ valueToRemoveFrom.value.slice(index + 1, valueToRemoveFrom.value.size)
           JsSuccess(JsArray(updatedJsArray))
         case valueToRemoveFrom: JsArray => JsError(s"array index out of bounds: \$index, \$valueToRemoveFrom")
-        case _ => JsError(s"cannot set an index on \$valueToRemoveFrom")
       }
     }
 
@@ -105,7 +104,7 @@ package object models {
 
           Reads.optionNoError(Reads.at[JsValue](JsPath(first :: Nil)))
             .reads(oldValue).flatMap {
-            opt: Option[JsValue] =>
+            (opt: Option[JsValue]) =>
 
               opt.map(JsSuccess(_)).getOrElse {
                 second match {