+++ src/main/g8/test/forms/behaviours/FieldBehaviours.scala
@@ -15,7 +15,7 @@ trait FieldBehaviours extends FormSpec with ScalaCheckPropertyChecks with Genera
     "bind valid data" in {
 
       forAll(validDataGenerator -> "validDataItem") {
-        dataItem: String =>
+        (dataItem: String) =>
           val result = form.bind(Map(fieldName -> dataItem)).apply(fieldName)
           result.value.value mustBe dataItem
           result.errors mustBe empty